renamed home sub-folders
