+++ lib/src/config/routes/app_router.dart
@@ -6,7 +6,7 @@ import 'package:irrigazione_iot/src/config/routes/go_router_refresh_stream.dart'
 import 'package:irrigazione_iot/src/features/authentication/data/auth_repository.dart';
 import 'package:irrigazione_iot/src/features/authentication/screen/sign_in/sign_in_screen.dart';
 import 'package:irrigazione_iot/src/features/dashboard/screen/dashboard_screen.dart';
-import 'package:irrigazione_iot/src/features/home/presentation/home_nested_navigator.dart';
+import 'package:irrigazione_iot/src/features/home/screen/home_nested_navigator.dart';
 import 'package:irrigazione_iot/src/features/pumps/presentation/add_pump/add_update_pump_form.dart';
 import 'package:irrigazione_iot/src/features/pumps/presentation/pump_details/pump_details_screen.dart';
 import 'package:irrigazione_iot/src/features/pumps/presentation/pump_list/pumps_list_screen.dart';
